+++ src/components/material-beautify-chinese-study/beautify-chinese-study.tsx
@@ -3,7 +3,6 @@ import { PhoneticType } from '../../enums/PhoneticType';
 import { HanziType } from '../../enums/HanziType';
 import Vowel from '../../phonetics/Vowel';
 import Zhuyin from '../../phonetics/Zhuyin';
-import HanziWriter from 'hanzi-writer';
 import * as OpenCC from 'opencc-js';
 import * as PinyinGenerator from 'pinyin';
 
@@ -87,9 +86,6 @@ export class MaterialBeautifyChineseStudy {
 	public generatedTraditional: string = '';
 	@State()
 	public generatedTraditionalSentence: string = '';
-
-	@Element()
-	private element: HTMLElement;
 
 	private conversionConfig: object = { from: 'cn', to: 'tw' };
 	private template: HTMLElement;
@@ -358,70 +354,6 @@ export class MaterialBeautifyChineseStudy {
 		return hanziToPhoneticCharacters(phonicType, phonicValue);
 	}
 
-	private createStrokeOrderCharacter(): void {
-		let delayBetweenAnimations: number = 500;
-		let characters: Array<string> = this.primaryCharacter.split('');
-		let drawingArea: HTMLElement = this.element.shadowRoot.querySelector('#stroke-order');
-		while (drawingArea.firstChild) {
-			drawingArea.removeChild(drawingArea.firstChild);
-		}
-		let strokeOrderCharacters: Array<any> = [];
-
-		for (var i = 0; i < characters.length; i++) {
-			let idName = 'character-' + i;
-			let node = document.createElement('div');
-			node.id = idName;
-			drawingArea.appendChild(node);
-
-			let strokeOrderCharacter = HanziWriter.create(
-				node,
-				characters[i], {
-					strokeColor: '#000000',
-					width: 50,
-					height: 50,
-					padding: 1,
-					delayBetweenStrokes: 500,
-				}
-			);
-			strokeOrderCharacters.push(strokeOrderCharacter);
-		}
-
-		const animateNextCharacter = (current: number, hanziCharacters: Array<any>) => {
-			const currentCharacter: any = hanziCharacters[current];
-			const nextId: number = current + 1;
-			const nextCharacter: any = hanziCharacters[nextId];
-			const nextNextId: number = current + 2;
-			const nextNextCharacter: any = hanziCharacters[nextNextId];
-			const isLastCharacter = () => nextCharacter === undefined;
-			const isNextCharacterTheLastCharacter = () => nextNextCharacter === undefined;
-
-			setTimeout(() => {
-				if (isLastCharacter()) {
-					currentCharacter.animateCharacter();
-				} else {
-					if (isNextCharacterTheLastCharacter()) {
-						nextCharacter.animateCharacter();
-					} else {
-						nextCharacter.animateCharacter({
-							onComplete: () => animateNextCharacter(nextId, hanziCharacters)
-						});
-					}
-				}
-			}, delayBetweenAnimations);
-		};
-
-		const animateFirstCharacterOnly = () => {
-			let firstCharacter: any = strokeOrderCharacters[0];	
-			firstCharacter.animateCharacter({
-				onComplete: () => animateNextCharacter(0, strokeOrderCharacters)
-			});
-		};
-
-		setTimeout(() => {
-			animateFirstCharacterOnly()
-		}, delayBetweenAnimations + delayBetweenAnimations);
-	};
-
 	private createCard(): HTMLElement {
 		const template: HTMLElement =
 			<div id='anki-background'>
@@ -464,14 +396,5 @@ export class MaterialBeautifyChineseStudy {
 				{ this.template }
 			</Host>
 		);
-	}
-
-	/**
-	 * Called after every `render()`.
-	 */
-	 public componentDidRender(): void {
-		if (this.getCardType() === 'writing' && this.getCardOrientation() === 'answer') {
-			this.createStrokeOrderCharacter();
-		}
 	}
 }
