+++ src/components/material-beautify-chinese-study/styles/card-type.css
@@ -27,6 +27,11 @@ p {
     background-color: #A8DADC;
 }
 
+:host(.writing) {
+    color: #5BC0BE;
+    background-color: #3A506B;
+}
+
 :host(.tones) {
     color: #FFD166;
     background-color: #EF476F;
