+++ src/components/material-beautify-chinese-study/styles/hanzi-with-phonic.css
@@ -1,7 +1,10 @@
-#vocabulary {
-  /* purposefully left undefined - default font size is fine  */
+#phonic-only > ruby > a > div {
+  display: none;
 }
 
+#phonic-only-phonic {
+  font-size: xx-large;
+}
 
 #primary-item {
   font-size: x-large;
