+++ src/index.html
@@ -23,7 +23,7 @@
 	<material-beautify-chinese-study id='asdf'
 		primary-character='天气'
 		primary-character-sentence='今天天气很好。'
-		card-type='audio'
+		card-type='writing'
 		meaning='weather'
 		sentence-meaning='The weather is nice today.'
 		card-orientation='answer'
